ophys and ephys device
