+++ src/aind_data_schema/ophys/ophys_rig.py
@@ -8,6 +8,8 @@ from typing import List, Optional
 
 from pydantic import BaseModel, Field
 
+from ..device import Device, DeviceType, DeviceManufacturer
+
 
 class CameraName(Enum):
     """Camera name"""
@@ -17,13 +19,10 @@ class CameraName(Enum):
     FACE_CAMERA = "Face Camera"
 
 
-class Camera(BaseModel):
+class Camera(Device):
     """Description of an ophys camera"""
 
     name: CameraName = Field(..., title="Camera Name")
-    manufacturer: str = Field(..., title="Manufacturer")
-    model: str = Field(..., title="Model")
-    serial_number: str = Field(..., title="Serial number")
     position_x: float = Field(..., title="Position X")
     position_y: float = Field(..., title="Position Y")
     position_z: float = Field(..., title="Position Z")
@@ -41,14 +40,6 @@ class CameraType(Enum):
 
     CAMERA = "Camera"
     PMT = "PMT"
-    OTHER = "other"
-
-
-class DetectorManufacturer(Enum):
-    """Detector manufacturer name"""
-
-    HAMAMATSU = "Hamamatsu"
-    PCOS = "PCOS"
     OTHER = "other"
 
 
@@ -76,7 +67,7 @@ class Immersion(Enum):
     OIL = "oil"
 
 
-class Detector(BaseModel):
+class Detector(Device):
     """Description of a generic detector"""
 
     name: Optional[str] = Field(
@@ -85,54 +76,9 @@ class Detector(BaseModel):
         title="Name",
     )
     type: CameraType = Field(..., title="Camera Type")
-    manufacturer: DetectorManufacturer = Field(
-        ..., title="Detector Manufacturer"
-    )
-    model: str = Field(..., title="Model")
-    serial_number: Optional[str] = Field(None, title="Serial number")
     data_interface: DataInterface = Field(..., title="Data interface")
     cooling: Cooling = Field(..., title="Cooling")
     immersion: Optional[Immersion] = Field(None, title="Immersion")
-
-
-class DeviceType(Enum):
-    """Device type name"""
-
-    DIFFUSER = "Diffuser"
-    GALVO = "Galvo"
-    BEAM_EXPANDER = "Beam expander"
-    LASER_COUPLER = "Laser coupler"
-    PRISM = "Prism"
-    OBJECTIVE = "Objective"
-    SLIT = "Slit"
-    OTHER = "Other"
-
-
-class DeviceManufacturer(Enum):
-    """Device manufacturer name"""
-
-    THORLABS = "Thorlabs"
-    OPTOTUNE = "Optotune"
-    CAMBRIDGE_TECHNOLOGY = "Cambridge Technology"
-    NIKON = "Nikon"
-    EDMUND_OPTICS = "Edmund Optics"
-    EALING = "Ealing"
-    HAMAMATSU = "Hamamatsu"
-    OTHER = "Other"
-
-
-class Device(BaseModel):
-    """Description of an ophys device"""
-
-    type: DeviceType = Field(
-        ...,
-        description="Type of device. If Other please describe in Notes.",
-        title="Type",
-    )
-    manufacturer: DeviceManufacturer = Field(..., title="Manufacturer")
-    model: Optional[str] = Field(None, title="Model")
-    serial_number: Optional[str] = Field(None, title="Serial number")
-    notes: Optional[str] = Field(None, title="Notes")
 
 
 class FilterType(Enum):
@@ -156,12 +102,10 @@ class FilterSize(Enum):
     FILTER_SIZE_32 = 32
 
 
-class Filter(BaseModel):
+class Filter(Device):
     """Description of a filter"""
 
     type: FilterType = Field(..., title="Filter Type")
-    manufacturer: FilterManufacturer = Field(..., title="Filter Manufacturer")
-    model: str = Field(..., title="Model")
     size: Optional[FilterSize] = Field(None, title="Size (mm)")
     cut_off_frequency: Optional[int] = Field(None, title="Cut off frequency")
     cut_on_frequency: Optional[int] = Field(None, title="Cut on frequency")
@@ -189,14 +133,11 @@ class Coupling(Enum):
     OTHER = "other"
 
 
-class Laser(BaseModel):
+class Laser(Device):
     """Description of a laser"""
 
     name: LaserName = Field(..., title="Laser Name")
-    manufacturer: str = Field(..., title="Manufacturer")
-    model: str = Field(..., title="Model")
     item_number: Optional[str] = Field(None, title="Item number")
-    serial_number: str = Field(..., title="Serial number")
     wavelength: int = Field(..., title="Wavelength (nm)")
     maximum_power: float = Field(..., title="Maximum power (mW)")
     coupling: Optional[Coupling] = Field(None, title="Coupling")
@@ -218,11 +159,9 @@ class LensSize(Enum):
     LENS_SIZE_2 = 2
 
 
-class Lens(BaseModel):
+class Lens(Device):
     """Description of a lens"""
 
-    manufacturer: Optional[str] = Field(None, title="Manufacturer")
-    model: Optional[str] = Field(None, title="Model")
     focal_length: Optional[float] = Field(None, title="Focal length (mm)")
     size: Optional[LensSize] = Field(None, title="Size (inches)")
     optimized_wavelength_range: Optional[str] = Field(
@@ -239,13 +178,10 @@ class PatchName(Enum):
     PATCH_CORD_C = "Patch Cord C"
 
 
-class Patch(BaseModel):
+class Patch(Device):
     """Description of a patch"""
 
     name: PatchName = Field(..., title="Patch Name")
-    manufacturer: Optional[str] = Field(None, title="Manufacturer")
-    part_number: str = Field(..., title="model")
-    serial_number: str = Field(..., title="Serial number")
     core_diameter: float = Field(..., title="Core diameter (um)")
     numerical_aperture: float = Field(..., title="Numerical aperture")
     photobleaching_date: Optional[date] = Field(
